create show_preview() for counter and pose
show_preview() takes a single preview (if the camera allows) and
displays it with an optional message superimposed.

show_counter now calls show_preview() in a loop.

(Note that show_counter no longer needs a special case if the camera
does not handle previews as show_preview will still correctly display
the countdown message on a black background.)