add show_pose() which shows live preview while saying "POSE"
Instead of having take_pictures() show the word POSE on a black
background and then sleep for 2 seconds, we now call a new function,
show_pose, that gets a live preview and superimposes the "POSE"
message. This allows people to actually pose.

As an added benefit, this lets V4L ramp the brightness correctly
before taking the picture, if we're using OpenCV and a webcam.
(Apparently, you need to read about 30 frames and discard them).