Write to shared memory instead of sd card if possible
Previously, images during the preview were written to /tmp/, which on
the Raspberry Pi is usually an SD card. This is not only slow, but can
cause SD card failures. (Flash technology has a limited number of
writes).

Now images are written to /dev/shm/, if it exists. (Falls back to /tmp/.)