Add debugging to measure FPS during preview() by hitting '1'
Latency and FPS are very bad, at least with my webcam. This makes it
hard to pose for a picture.

I'm not sure where the bottleneck is, but by doing the conversion from
OpenCV to PyGame in memory instead of writing files, I'm able to
improve both latency and FPS dramatically.

Currently this code will benchmark the original method if True and the
cv2/numpy method if False.