Make debug code cleaner, closer to integration in mainline
show_preview_fps_2() cleaned up (and sped up) so that it uses the
interface offered by gui.py as much as possible instead of writing
directly using PyGame.

For some reason make_surface() is running only 8fps on the iMac I'm on
(it was much faster on the Raspberry Pi), but I can work around it by
creating a new surface and blitting into that (15fps). (THIS IS WEIRD!)

I also copied the scaling code from the gui.py so that the image is
scaled proportionately.