Add functions to save preview directly to memory instead of a file
Two functions added:

get_preview_array(), which returns a Numpy array.

get_preview_pygame_surface(), which returns a Pygame Surface.

Note that one can also specify a "maximum size" to these functions.
The returned array will be decimated to be no more than the max size.
This makes it easy later to blit directly to the display without
needing to use an expensive Pygame.transform.scale(). It is faster to
do the decimation here, because, inside these functions, we can
perform it before the other transformations (rotation, color convert).

The downside of using this "maximum size" feature is that the preview
image will not completely fill the screen. On slow hardware, like the
raspberry pi, it's (probably) worth it for the extra smoothness of the
preview.