Print opencv camera capabilities, clean up code
Previous code was using uncommented magic numbers:

-            self.cap.set(3, picture_size[0])
-            self.cap.set(4, picture_size[1])

I had assumed that this was requesting the width and height of the
video capture frame, but oddly the picture_size being passed in from
photobooth.py was the maximum assembled image size, which is way too
large. I would have guessed it would have used the "thumb_size".

After investigation, I'm guessing that this was either a bug in the
previous code or an attempt to force the camera to the highest
resolution possible. (Requesting a resolution too high will cause
OpenCV to fall back to the next best available.)

It doesn't make much sense to request a higher resolution than you'll
be using as it only slows the camera down. So, I'll fix that in an
upcoming patch, for now, I'll just add in the cleanup I did to make
the code more understandable.