Cache rendered_text since it is so darn slow
Added set_ and get_message_cache() for use by show_message().

On a raspberry pi 3b, the show_count() routine was extremely slow
because it was trying to render text (the countdown numbers) as fast
as the framerate of the camera. This made the camera preview lag
badly.

Perhaps show_count ought to be smarter about not doing that, but I
felt like it made more sense to just cache all rendered text (we don't
actually have that many messages we show) and then return it
immediately.