Add error checking for OpenCV cameras
Previously, if there was no camera plugged in, photobooth.py would
simply say, "SERIOUS ERROR". Now, it will inform you of the problem
and, if you plug in a camera, it will initialize it and work properly.

Note that due to a bug in OpenCV 2.9.4, a webcam that has problems
AFTER it has been opened successfully (e.g., gets unplugged while
photobooth is running) will NOT return correct error codes. That is,
the return value from cap.read() is always True, so we have no way of
knowing it needs to be reinitialized.