Use wall time, not CPU time for delays
Previously, the photobooth called time.clock() to figure out how many
seconds had elapsed. This is a bad idea as it is dependent on the
amount of CPU used, not real time. This was noticeable in how, during
the preview, the countdown felt slower than acceptable and non-uniform.

Simply changing the call to time.time() fixed the problem.