Reinitialize camera before POSE, remove profiling code
The OpenCV webcam slowdown can be worked around by reopening the
camera. Now the camera is always reinitialized after the user presses
the button. (Actually, we do it after showing the word "POSE", since
initialization takes a second, we want something on the screen so the
user knows the program is working).

Also added a new key, 'i' to force reinitialization. This could also
be handy if the web driver gets confused and you want to unplug/replug
the webcam without quitting the program.

Now that we know what the slowdown is, I've removed the calls to
{begin,end}_profile(). However, I've left the functions themselves in
place as they may come in handy in the future.