Remove global thumb_size variable
The original code had an unused variable named "thumb_size". I'm not
sure what its purpose was supposed to be. Perhaps it was intended to
override the calculated thumb_size in assemble_pictures. If so, it
will be easy enough to add back in if ever that functionality is
implemented.