Reset default aspect ratio to 6x4
max_assembled_size is now 1440x960

max_assembled_size used to be 2352x1568, which was perfect for
printing to 4x6 photo paper. However, it left huge white margins
around my low-res (640x480) camera.

To fix the margins I had initially used (1280, 960) == 2 * 640x480 as
the assembled_size, but that meant printing was ambiguous since some
printers would center the image and some wouldn't. To fix that --- at
least for people printing to 4x6 photo paper --- the default is now
(1440, 960).

Ideally, this ought to be dynamically computed or something.