Cleanup assemble_pictures(), make {inner,outer}_border dynamic
Previously, the white border was hardcoded as 50 pixels outside the
thumbnails and 20 pixels in between. This didn't make sense when
printing at different resolutions.

The outer border is now 2% of the width (or the height if it is
portrait instead of landscape). The inner border is 1%.

For example, if the assembled_size is (2352, 1568), which was the
previously hardcoded value, then the borders will be 48 and 24. This
seems reasonably close to the old 50 and 20.