Resize thumbnails larger when assembling
Use PIL's Image.resize() instead of Image.thumbnail(). Also add
maxpect() function for calculating the maximum aspect ratio.

Previously, images in the montage were assumed to be larger than
thumb_size and only scaled down, never scaled up. This left too much
white space when using a lousy webcam like mine.

The previous fix for this was to reduce the assembled_size, but that
seemed clunky since it is camera dependent. Now that thumbnails can be
scaled up as well as down, I've reset assembled_size to 2352x1568, the
same as the original author of this code had it. (Why she/he choose
that number, I'm not sure. It appears to be 4"x6" photo paper at 392 dpi).