Remove unnecessary sudo
Running photobooth.py as root was causing files to be owned by root
instead of 'pi'. It seems unnecessary on my Raspian Pi, as it runs
fine without sudo (at least the OpenCV module). I can't completely
test gphoto, but that seems to work as well.

(Running as sudo is also considered unhygenic.)