Add error checking and handling for gvfs-gphoto problems.
In the README, we explain to people that they need to do weird things
to disable gvfs (like delete files that will simply reappear next time
they update their system). That is ugly and there is a better way:
hide the ugliness in a shell script! :-)

The photobooth.sh wrapper script now checks if gphoto2 can detect a
camera but cannot reset the USB bus, which is a sign that some other
process has control of it. The script then goes through some
machinations to try to get gphoto working, including using:

    gsettings set org.gnome.desktop.media-handling automount false

and

    killall gvfsd-gphoto2 gvfs-gphoto2-volume-monitor

The first command tells the system not to mount the camera when it's
connected in the future. However, it is likely already mounted and we
don't want to make people unplug the camera and start over, especially
on a kiosk device. So, the second command kills the most common
culprits so we can, hopefully, forge ahead.