Fix piggyphoto previews & comment out viewfinder
Piggyphoto wasn't working because I forgot to turn the result of the
preview from a PIL Image into a Numpy array.

There was something from br's original code called set_idle() which
was trying to do something to the viewfinder, but it was just causing
my python to bomb out with no such such-and-such. I commented it out
and all was well.