Implement preview_rotate, fix tmp_dir problem
Some cameras (e.g., Canon PowerShot A510) have a gravity sensor so a
captured image is correctly rotated if the camera is on its side, but
preview images are NOT. Now you can set preview_rotate=True when
initializing Camera_gPhoto() if your camera is like that.

Also, removed hardcoding of the "/tmp" directory. Now, we use
/dev/shm, if it exists, and fallback to /tmp.