Autodetect gphoto2cffi capture bug
By default, gphoto2cffi is supposed to capture from RAM without saving
to SD card. However, it is buggy, at least on my Canon A510 and
throws a "File not found" exception after taking the picture.

Added code that detects when that happens and switches to backup code
that manually saves the image to the camera's SD card, downloads it,
and deletes it.