gphoto2cffi should be optional. Fix gphoto2 cmdline "not found"
Accidentally was trying to except on error that wouldn't exist if
gphoto2cffi didn't get imported. Placed it in the 'if' where it
belongs.

Also, if cffi and piggyphoto didn't exist, camera.py would run
'gphoto2' as a subprocess. However, if the command didn't exist, it
wouldn't know it because it looked for "command not found" (which is
what bash says) instead of "gphoto2: not found" (which is what dash,
the /bin/sh on Debian machines uses). Fixed now.