Show an error if CFFI's get_preview() fails
Sometimes my Canon A510 returns an empty string from get_preview() and
I have to turn off the camera and turn it back on. I'm not quite sure
where the bug is, I suspect it is gphoto2cffi, but in the meantime,
this patch will at least catch the bug and report it.