Workaround gphoto2 2.5.10's "thumb_" bug
Note that when using --capture-preview, the filename will *always* be
prefixed with "thumb_" as of gphoto2 2.5.10. This is very annoying. My
code for get_preview_array already checked for both filenames.
However, we needed a different solution for take_preview(). I ended up
adding a "verbose" flag to call_gphoto2() to find out what gphoto2 has
actually saved the preview file as, then using os.rename() to fix it.

Gross. This is clearly a bug in gphoto2.