A shell script for resetting *all* USB devices.
Useful for debugging since some gphoto2 operations seem to lockup the
firmware on my camera.

Theoretically, gphoto2 --reset should work the same, but it does not.
This can be seen by the fact that the camera lens closes when this
script is run, but, at least on my camera, gphoto2 --reset, leaves the
lens open.